MetaDome 10-20 pattern database
The MetaDome is the foundational data set that defines the electrode pattern used for the Ultracortex. It consists of a library of 3D scans that have been processed in a consistent way. A parametric division of the cranium based on anatomical landmarks is used to accurately define electrode locations ("nodes") for each subject, then these location are averaged together to create a generic pattern, suitable for general use. This is the second version of the MetaDome. The number of subjects is currently at 8, and represents a diverse sampling to create the best generic fit possible.